CKKS : new tests and benchmarks. RING : Ziggurat Sampling
